backend/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-environ==0.11.2
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.1
#psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.5.1
coreapi==2.3.3
coreschema==0.0.4
virtualenv==20.25.1
django-filter==24.2
setuptools>=65.5.0
django-cors-headers==4.3.1
djangorestframework-bulk==0.2.1
django-pandas==0.6.2
openpyxl == 3.1.0
